backend/ml-services/requirements.txt
flask
torch
torchvision
pillow
opencv-python
werkzeug
numpy
